Use separate canvases for forward/backward slices
This way only the forward slices, which are drawn underneath the backward ones, have margins. As a result we have tidier edges.
Also, we can now have zero points, which is just the straight image rotating around without any mirrors.
Also we now align the top with the edge of the first slice, rather than the middle of it.